+++ ORU2DICOM/Program.cs
@@ -1,0 +1,9 @@
+namespace DICOM7.ORU2DICOM
+{
+  internal class Program
+  {
+    public static void Main(string[] args)
+    {
+    }
+  }
+}